enable some tests for DB2 after Vert.x update
See #295
